rule_based_approach/requirements.txt
PyMuPDF
pyyaml
pytesseract
Pillow
